I belelieve my database got deleted
I think my database cluster got deleted for some reason. I have made some changes to the code and it should be working but I cannot tell for sure until I reconnect to the database. However, the app should allow the user to type in the info for a post and send the data to the database.